Updated formatting section
I think it is not a duplicate function but an added value to the existing functions. In my experience, the analysis of plots with only f__ "empty" as information is not useful and having f__Taxonomy_OTUID is convenient. This also helps in the plotting of networks from your tutorial.  I will add a new issue for this.